fixed macosMajorVersionLessThan to use float64
There are APIs which are supported above major and minor versions. like https://developer.apple.com/documentation/virtualization/vzvirtioblockdeviceconfiguration/3917717-blockdeviceidentifier?language=objc

So I fixed this API to check also minor version.

+++ virtualization_export_test.go
@@ -1,5 +1,5 @@
 package vz
 
-func MacosMajorVersionLessThan(version int) bool {
+func MacosMajorVersionLessThan(version float64) bool {
 	return macosMajorVersionLessThan(version)
 }
